инфа о выезде

+++ app/src/main/java/ru/zavbus/zavbusexample/entities/TripRecord.kt
@@ -20,7 +20,7 @@ data class TripRecord(
         @ColumnInfo(name = "name") val name: String,
         @ColumnInfo(name = "commentFromVk") val commentFromVk: String?,
         @ColumnInfo(name = "orderedKit") val orderedKit: String?,
-        @ColumnInfo(name = "prepaidSum") val prepaidSum: Long?,
+        @ColumnInfo(name = "prepaidSum") val prepaidSum: Int?,
         @ColumnInfo(name = "paidSumInBus") var paidSumInBus: Int?,
         @ColumnInfo(name = "confirmed") var confirmed: Boolean?,
         @ColumnInfo(name = "packetId") var packetId: Long,
